Move js packing functions to util.fiddling
