+++ CHANGELOG.md
@@ -66,7 +66,7 @@ The table below shows which release corresponds to each branch, and what date th
 - [#1922][1922] Fix logic in `wait_for_debugger`
 - [#1828][1828] libcdb: Load debug info and unstrip libc binary
 - [#1939][1939] Fix error in validating log levels
-- [#1979][1979] Add `js_escape()` and `js_unescape()` to `util.packing`
+- [#1979][1979] Add `js_escape()` and `js_unescape()` to `util.fiddling`
 
 [1922]: https://github.com/Gallopsled/pwntools/pull/1922
 [1828]: https://github.com/Gallopsled/pwntools/pull/1828
